\docType{package}
\name{funphylocom}
\alias{funphylocom}
\alias{funphylocom-package}
\alias{package-funphylocom}
\title{Functional traits, phylogenies, communities, simulations}
\description{
  Simulates the influence of functional traits and
  phylogenies on ecological communities, via the
  \code{\link{fpcomSims}} function.
}

