\name{plot.traitgram}
\alias{plot.traitgram}
\title{Plot traitgram objects}
\usage{
  \method{plot}{traitgram} (x, ...)
}
\arguments{
  \item{x}{A traitgram object}

  \item{...}{Arguments to pass to \code{\link{traitgram}}}
}
\value{
  No return value, plot produced
}
\description{
  Method for plotting objects of class traitgram
}

