\name{tail.list}
\alias{tail.list}
\title{Tail of a list}
\usage{
  \method{tail}{list} (x, n = 6L, ...)
}
\arguments{
  \item{x}{A list}

  \item{n}{Number of lines}

  \item{...}{Not used}
}
\value{
  The first \code{n} elements of each list element.
}
\description{
  utility function to make extractions of the trait values
  at the tips easier
}

